Updates setup, adds install

+++ setup.rb
@@ -6,30 +6,21 @@ class Setup
   def initialize
     prompt_for_neovim
     prompt_for_dotfiles
-    rbenv_rehash
-    install_pip_gems_for_neovim
+    set_permissions_for_install_script
     install_powerline_fonts
     goodbye
+  end
+
+  def set_permissions_for_install_script
+    `chmod +x ~/.dotfiles/install.sh`
   end
 
   def symlink_bash_profile
     `ln -s ~/.dotfiles/bash/bash_profile.sh ~/.bash_profile`
   end
 
-  def rbenv_rehash
-    `rbenv rehash`
-  end
-
-  def install_pip_gems_for_neovim
-    puts "Installing pip2 for neovim..."
-    `pip2 install neovim`
-    puts "Installing pip3 for neovim..."
-    `pip3 install neovim`
-    puts "Installing ruby gem for neovim..."
-    `sudo gem install neovim`
-  end
-
   def symlink_pgclirc
+    `mkdir ./config/pgcli/`
     `ln -s ~/.dotfiles/dotfiles/config ~/.config/pgcli/config`
   end
 
@@ -39,15 +30,9 @@ class Setup
     symlink_pgclirc
   end
 
-  def symlink_colors_for_neovim
-    `rm -rf /usr/local/Cellar/neovim/0.2.0_1/share/nvim/runtime/colors`
-    `ln -s ~/.dotfiles/neovim/colors /usr/local/Cellar/neovim/0.2.0_1/share/nvim/runtime/colors`
-  end
-
   def symlink_init_for_neovim
-    `mkdir ~/.config/`
-    `mkdir ~/.config/nvim/`
-    `ln -s ~/.dotfiles/neovim/init.vim ~/.config/nvim/init.vim`
+    `rm -rf /usr/local/Cellar/neovim/0.2.0_1/share/nvim/runtime/colors`
+    `ln -s ~/.dotfiles/nvim ~/.config/`
   end
 
   def prompt_for_neovim
@@ -61,11 +46,20 @@ class Setup
     gets.chomp.downcase
   end
 
+  def prompt_for_powerline
+    puts 'Would you like to install powerline fonts? (y/N)'
+    print '>> '
+    gets.chomp.downcase
+  end
+
   def install_powerline_fonts
-    `git clone https://github.com/powerline/fonts.git --depth=1`
-    `cd fonts`
-    `./install.sh`
-    `cd .. && rm -rf fonts`
+    response = prompt_for_powerline
+    if response.eql?('y')
+      `git clone https://github.com/powerline/fonts.git --depth=1`
+      `cd fonts`
+      `./install.sh`
+      `cd .. && rm -rf fonts`
+    end
   end
 
   def goodbye
